app/requirements.txt
asgiref==3.3.2
Django==3.2.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
djangorestframework
